Changed reading of csv file to use CsvHelper library and added src folder Co-authored-by: My <mmga@itu.dk>  Co-authored-by: Louis <lofk@itu.dk>  Co-authored-by: Kris <krme@itu.dk>  Co-authored-by: Hassan <halw@itu.dk>
